Update bot.js
Please test this one. I was not able to set up a test environment for the bot. I think it would improve visibility and debugging if we get the most current value directly from the database. 
I will not create a PR yet on this one because I do not know whether the extra clutter is worth it.